Huge refactor. Adds zoom ranges to source selections.
A lot of the code for downloading and processing tiles was
duplicated in each source. This has now been factored out into the
command process. The downloads are now all run through a single
queue, which allows them to all be run in parallel, which helps
when a small number of requests (e.g: GMTED) take a
disproportionate amount of time to download (and retry, retry,
retry...)

Also the source bounding boxes are now based on the padded
extents of the requested regions, and data sources now cache their
indexes.